requirements.txt
openai
pillow
bs4
requests
